Revise validation

+++ scripts/dl.py
@@ -42,9 +42,9 @@ def dl_flood_layers():
         filename = "{}".format(filename)
         out_path = os.path.join(folder, filename)
 
-        #if not os.path.exists(out_path):
+        if not os.path.exists(out_path):
 
-        urllib.request.urlretrieve(in_path, out_path)
+            urllib.request.urlretrieve(in_path, out_path)
 
     return
